Move VehicleAirPlacementArgs to VehiclePlaceAir.h
