+++ src/OpenLoco/src/Windows/Vehicle.cpp
@@ -16,6 +16,7 @@
 #include "GameCommands/Vehicles/VehiclePickupAir.h"
 #include "GameCommands/Vehicles/VehiclePickupWater.h"
 #include "GameCommands/Vehicles/VehiclePlace.h"
+#include "GameCommands/Vehicles/VehiclePlaceAir.h"
 #include "GameCommands/Vehicles/VehicleRearrange.h"
 #include "GameCommands/Vehicles/VehicleRefit.h"
 #include "GameCommands/Vehicles/VehicleReverse.h"
